task_manager/requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
kombu==4.6.11
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
vine==1.3.0
